requirements.txt
httpx~=0.23.0
starlette==0.20.4
fastapi==0.85.0
pydantic~=1.10.2
uvicorn~=0.18.3
python-dotenv~=0.21.0
httpx[http2]~=0.23.0
Jinja2~=3.1.2
rollbar~=0.16.3